+++ src/main/java/nimbus/exceptions/NimbusException.java
@@ -1,3 +1,4 @@
+package nimbus.exceptions;
 public class NimbusException extends Exception {
     public NimbusException(String message) {
         super(message);
