+++ src/main/java/nimbus/tasks/Deadline.java
@@ -1,12 +1,14 @@
+package nimbus.tasks;
+
+import nimbus.exceptions.NimbusException;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
 import java.util.Arrays;
 import java.util.List;
 
-public class Event extends Task {
-    private LocalDateTime from;
-    private LocalDateTime to;
+public class Deadline extends Task {
+    private LocalDateTime by;
     private static final List<DateTimeFormatter> INPUT_FORMATS = Arrays.asList(
             DateTimeFormatter.ofPattern("yyyy-MM-dd HHmm"),
             DateTimeFormatter.ofPattern("dd/MM/yyyy HHmm"),
@@ -15,10 +17,9 @@ public class Event extends Task {
     );
     private static final DateTimeFormatter OUTPUT_FORMAT = DateTimeFormatter.ofPattern("MMM dd yyyy, h:mm a");
 
-    public Event(String description, String from, String to) throws NimbusException {
+    public Deadline(String description, String by) throws NimbusException {
         super(description);
-        this.from = parseDateTime(from);
-        this.to = parseDateTime(to);
+        this.by = parseDateTime(by);
     }
 
     private LocalDateTime parseDateTime(String dateTimeString) throws NimbusException {
@@ -37,15 +38,15 @@ public class Event extends Task {
 
     @Override
     public String toString() {
-        return "[E]" + super.toString() + " (from: " + from.format(OUTPUT_FORMAT) + " to: " + to.format(OUTPUT_FORMAT) + ")";
+        return "[D]" + super.toString() + " (by: " + by.format(OUTPUT_FORMAT) + ")";
     }
 
     @Override
     public String toFileString() {
-        return "E | " + (isDone ? "1" : "0") + " | " + description + " | " + from.format(INPUT_FORMATS.get(0)) + " | " + to.format(INPUT_FORMATS.get(0));
+        return "D | " + (isDone ? "1" : "0") + " | " + description + " | " + by.format(INPUT_FORMATS.get(0));
     }
 
     public boolean isOnDate(LocalDateTime date) {
-        return from.toLocalDate().equals(date.toLocalDate()) || to.toLocalDate().equals(date.toLocalDate());
+        return by.toLocalDate().equals(date.toLocalDate());
     }
 }
