+++ src/main/java/nimbus/tasks/Task.java
@@ -1,3 +1,5 @@
+package nimbus.tasks;
+
 public abstract class Task {
     protected String description;
     protected boolean isDone;
