+++ src/main/java/nimbus/tasks/Todo.java
@@ -1,3 +1,5 @@
+package nimbus.tasks;
+
 public class Todo extends Task {
     public Todo(String description) {
         super(description);
