+++ src/main/java/nimbus/ui/UI.java
@@ -1,3 +1,8 @@
+package nimbus.ui;
+
+import nimbus.tasks.Task;
+import nimbus.tasks.Deadline;
+import nimbus.tasks.Event;
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 import java.util.ArrayList;
